+++ lib/src/apollovm_base.dart
@@ -33,7 +33,7 @@ import 'languages/wasm/wasm_generator.dart';
 /// The Apollo VM.
 class ApolloVM implements VMTypeResolver {
   // ignore: non_constant_identifier_names
-  static final String VERSION = '0.0.41';
+  static final String VERSION = '0.0.42';
 
   static int _idCount = 0;
 
@@ -133,7 +133,7 @@ class ApolloVM implements VMTypeResolver {
   }
 
   /// Creates a runner for the [language].
-  ApolloLanguageRunner? createRunner(String language) {
+  ApolloRunner? createRunner(String language) {
     switch (language) {
       case 'dart':
         return ApolloRunnerDart(this);